backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# ML & Data
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==2.0.3
shap==0.43.0

# Utilities
requests>=2.31.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
cachetools>=5.3.0
email-validator>=2.0.0
python-dotenv>=1.0.0
gunicorn==21.2.0
psutil>=5.9.0

# Build Tools
setuptools>=68.0.0
wheel>=0.42.0
pip>=23.3.0

# Model Storage (HuggingFace Hub)
huggingface-hub>=0.19.0

# Image & QR Processing
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# Document Parsing
python-docx>=1.0.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
httpx>=0.25.0
